Fixed large values bugs -> BigDecimal, fixed requerying when changing the period interval, added titles for the charts, fixed ordering of the countries in the charts because of the multithreading

+++ src/main/java/view/DataDisplayWrapper.java
@@ -11,6 +11,7 @@ import javafx.scene.layout.HBox;
 import javafx.stage.Stage;
 import table_view.TableViewPane;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
@@ -39,11 +40,11 @@ public class DataDisplayWrapper extends Stage {
     }
 
 
-    public void setData(ArrayList<TreeMap<Integer, Double>> inData) {
+    public void setData(ArrayList<TreeMap<Integer, BigDecimal>> inData) {
         data.clear();
-        for (TreeMap<Integer, Double> val : inData) {
+        for (TreeMap<Integer, BigDecimal> val : inData) {
             TreeMap<Integer, Number> toAdd = new TreeMap<>();
-            for(Map.Entry<Integer,Double> entry : val.entrySet()) {
+            for(Map.Entry<Integer,BigDecimal> entry : val.entrySet()) {
                 toAdd.put(entry.getKey(), entry.getValue());
             }
             data.add(toAdd);
